Separate CallToActionItem and move to layouts folder
