--- code/src/pages/home/components/call-to-action-section/messages/index.ts
@@ -1,4 +1,0 @@
-import en from "@/pages/home/components/call-to-action-section/messages/en.json";
-import uk from "@/pages/home/components/call-to-action-section/messages/uk.json";
-
-export default { en, uk };
